eventbridge-pipes-dynamodb-to-eventbridge-cdk/requirements.txt
aws-cdk-lib>=2.173.2
constructs>=10.0.0,<11.0.0
pytest
boto3
